final comit

+++ MyDll.h
@@ -1,5 +1,5 @@
 #define EXPORT __declspec(dllexport)
 
-EXPORT double asciiToDouble(char *);
+EXPORT float asciiToDouble(char *);
 EXPORT int asciiHEXToInt(char *);
 EXPORT int asciiBinaryToInt(char *);
